Deprecate root webhook query
